% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_ploidy.R
\name{adjust_ploidy}
\alias{adjust_ploidy}
\title{Adjust ploidy for samples with CNV data.}
\usage{
adjust_ploidy(
  this_seg,
  seg_path,
  projection = "grch37",
  pga,
  pga_cutoff = 0.05,
  exclude_sex = TRUE,
  return_seg = TRUE
)
}
\arguments{
\item{this_seg}{Input data frame of seg file.}

\item{seg_path}{Optionally, specify the path to a local seg file.}

\item{projection}{Argument specifying the projection of seg file, which will determine chr prefix and genome size. Default is grch37, but hg38 is also accepted.}

\item{pga}{If PGA is calculated through other sources, the data frame with columns sample_id and PGA can be provided in this argument.}

\item{pga_cutoff}{Minimum PGA for the sample to adjust ploidy. Default is 0.05 (5 \%).}

\item{exclude_sex}{Boolean argument specifying whether to exclude sex chromosomes from calculation. Default is TRUE.}

\item{return_seg}{Boolean argument specifying whether to return a data frame in seg-consistent format, or a raw data frame with all step-by-step transformations. Default is TRUE.}
}
\value{
A data frame in seg-consistent format with ploidy-adjusted log ratios.
}
\description{
\code{adjust_ploidy} returns a seg file with log.ratios adjusted to the overall sample ploidy.
}
\details{
This function adjusts the ploidy of the sample using the percent of genome altered (PGA). The PGA is calculated internally, but can also be optionally provided as data frame
if calculated from other sources. Only the samples above the threshold-provided PGA will have ploidy adjusted. The function can work with either individual or
multi-sample seg file. The telomeres are always excluded from calculation, and sex chromosomes can be optionally included or excluded. The supported projections are grch37 and hg38.
The chromosome prefix is handled internally per projection and does not need to be consistent.
}
\examples{
\dontrun{
all_meta <- GAMBLR.open::get_gambl_metadata() 
all_sample_seg = GAMBLR.open::get_cn_segments(these_samples_metadata = all_meta)
all_sample_seg = dplyr::rename(all_sample_seg, "sample" = "ID")

sample_seg = dplyr::filter(all_sample_seg, sample == "DOHH-2")
adjust_ploidy(this_seg = sample_seg)
}


}
