% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_igh_breakpoints.R
\name{annotate_igh_breakpoints}
\alias{annotate_igh_breakpoints}
\title{Annotate IGH Breakpoints.}
\usage{
annotate_igh_breakpoints(annotated_df, genome_build = "grch37")
}
\arguments{
\item{annotated_df}{Previously annotated data frame of SVs.}

\item{genome_build}{Version of reference build to be used, only grch37 currently accepted.}
}
\value{
A slightly modified bedpe with added columns.
}
\description{
Add annotation to IGH breakpoints to infer mechanism based on location within IGH.
}
\details{
Returns a modified bedpe with additional columns.
The most useful columns that are added are mechanism (one of CSR, AID, VDJ) and label (NA if unmatched, otherwise one of Emu, Smu, one of the J segments or switch regions).
}
\examples{
\dontrun{
library(GAMBLR.open)

# Get Manta SVs
my_meta = GAMBLR.open::get_gambl_metadata()
my_samples = my_meta$sample_id
manta_sv = get_manta_sv(these_samples_metadata = my_meta,
                        these_sample_ids = my_samples, verbose = FALSE)

# Annotate IGH breakpoints
all_annotated = annotate_sv(sv_data = manta_sv)
ig_annotated = annotate_igh_breakpoints(all_annotated)
}
}
