% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_intrachromosomal_sv.R
\name{annotate_intrachromosomal_sv}
\alias{annotate_intrachromosomal_sv}
\title{Annotate intrachromosomal SV.}
\usage{
annotate_intrachromosomal_sv(
  sv_data,
  return_as = "bedpe",
  projection = "grch37",
  types = c("DEL", "DUP", "INS", "INV", NA),
  max_size = 1e+05,
  keep_genes = "all",
  verbose = FALSE
)
}
\arguments{
\item{sv_data}{A data frame of SVs. This should be the output of get_manta_sv.
The required columns are columns for both chromosomes, coordinates and strand plus SOMATIC_SCORE and tumour_sample_id.}

\item{return_as}{Stated format for returned output, default is "bedpe". Other accepted output formats are "bed" and "bedpe_entrez" (to keep entrez_ids for compatibility with portal.R and cBioPortal).}

\item{projection}{Reference genome build parameter. Can be one of grch37 or hg38. The default is grch37.}

\item{types}{A vector of sv types to be annotated. The default is set to "DEL", "DUP", "INS", and "INV".}

\item{max_size}{An integer specifying the maximum distance between start of first breakpoint and end of second breakpoint. The default is 100000 (100000 bp).}

\item{keep_genes}{Optional argument allowing ot only return svs annotated to the list of genes provided in this argument. The default is "all" (return all svs annotated to all genes).}

\item{verbose}{Optionally control whether to print any status info to console. Useful for debugging. Default is FALSE (no messages).}
}
\value{
A data frame with annotated SVs (gene symbol and entrez ID).
}
\description{
Annotate the intrachromosomal events in a data frame of SV breakpoints.
}
\details{
Specify a data frame with SVs (preferably the output from get_manta_sv() to the \code{sv_data} parameter and get back the same data frame with SV annotations.
}
\examples{
sv_df = get_manta_sv(verbose = FALSE)
annotated = annotate_intrachromosomal_sv(sv_data = sv_df)

}
