% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_browser_hub.R
\name{build_browser_hub}
\alias{build_browser_hub}
\title{Build UCSC browser track hub}
\usage{
build_browser_hub(
  regions_bed = GAMBLR.data::grch37_ashm_regions,
  these_sample_ids = NULL,
  these_samples_metadata = NULL,
  these_seq_types = c("genome", "capture"),
  projection = "grch37",
  local_web_host_dir = NULL,
  hub_dir = "my_hub",
  as_bigbed = TRUE,
  splitColumnName = "pathology",
  hub_name = basename(hub_dir),
  shortLabel = basename(hub_dir),
  longLabel = basename(hub_dir),
  contact_email,
  visibility = "squish",
  bigDataUrl_base = "https://github.com/morinlab/LLMPP/blob/main"
)
}
\arguments{
\item{regions_bed}{A BED-format table with the regions you want to retrieve SSMs
from. The columns 1, 2 and 3 must be chromosome names, start positions and
end positions, respectively. The default is the \code{GAMBLR.data::grch37_ashm_regions}
object which contains public-available SSMs from aSHM regions, but any other
regions can be provided.}

\item{these_sample_ids}{A vector of sample IDs that you want results for.}

\item{these_samples_metadata}{A metadata table (with sample IDs in a column) to
subset the samples of interest.}

\item{these_seq_types}{A vector of one or more seq types you want results for.
Possible values are "genome" or "capture". If more than
one seq type is provided, for each value in \code{splitColumnName}, the function
creates a separate track for each provided seq type. The default is
\code{c("genome", "capture")}. See the \strong{Details} section for more information.}

\item{projection}{The projection genome build. One of "grch37" (default) or "hg38".}

\item{local_web_host_dir}{Path to the directory where is your local copy of the
web host to be used. For example, if the hub should be hosted on GitHub,
\code{local_web_host_dir} should be the path to your local copy of the repository
directory. Default is NULL.}

\item{hub_dir}{Path to the directory (inside a web host) where you want to build
your track hub. If this directory does not exist, it will be created.}

\item{as_bigbed}{A Boolean value. If TRUE (default), custom tracks are saved as
bigBed (.bb) files. BED files otherwise.}

\item{splitColumnName}{An single string to indicate which metadata column is used
to split the MAF data into custom track files. Default is "pathology".}

\item{hub_name}{A string with the hub name (without spaces) to fill in the \code{hub}
field of the hub.txt file. Default is \code{basename(hub_dir)}.}

\item{shortLabel}{A string with the short hub label (maximum of 17 characters
recommended; spaces are allowed) to fill in the \code{shortLabel} field of the
hub.txt file. Default is \code{basename(hub_dir)}.}

\item{longLabel}{A string with the long hub label (maximum of 80 characters
recommended; spaces are allowed) to fill in the \code{longLabel} field of the
hub.txt file. Default is \code{basename(hub_dir)}.}

\item{contact_email}{Required parameter. A string with the contact email to fill
in the \code{email} field of the hub.txt file.}

\item{visibility}{A string that controls the track visibility mode. Possible
values are "pack", "dense", "full", or "squish" (default).}

\item{bigDataUrl_base}{A string with the base path of the web location of the
tracks' data files (\code{bigDataUrl} fields from the hub.txt file). For example,
if the hub should be hosted on GitHub, \code{bigDataUrl_base} should be something
like "https://github.com/morinlab/LLMPP/blob/main" (the default). See how the
track's paths from the bigDataUrl fields are set in the \strong{Details} section.}
}
\value{
Nothing.
}
\description{
Create a directory that contains the track hub files. They are:
a hub.txt file (named as \verb{<projection>_hub.txt} and marked with \verb{useOneFile on}),
and a subdirectory (named as the projection in use) with the custom tracks to
be visualized in UCSC browser. The custom track \code{regions} contains the regions
from where SSMs were retrieved. All other custom tracks contain SSMs from samples
separated by the \code{splitColumnName} parameter, where each file is named according
to the value in \code{splitColumnName} that it refers to.
}
\details{
\code{build_browser_hub} will create a custom track file for each combination of
\code{these_seq_types} and \code{splitColumnName} (if mutations can be found). Custom
track files are named as \verb{<a_seq_type_value>_<a_splitColumnName_value>.bb/bed}.

The \code{bigDataUrl} field of a track in the hub.txt file is defined in
the following way:

\if{html}{\out{<div class="sourceCode">}}\preformatted{file.path(bigDataUrl_base, hub_dir, projection, 
          paste0(track_file_names[i], "?raw=true\\n"))
}\if{html}{\out{</div>}}

where \code{track_file_names[i]} is the custom track file name.
}
\examples{
\dontrun{
# create a track hub in LLMPP GitHub repo

library(GAMBLR.data)

local_web_host_dir = "~/repos/LLMPP"
hub_dir = "hubs/ashm_test"

my_meta = get_gambl_metadata() \%>\% 
  filter(pathology \%in\% c("BL", "DLBCL", "FL"))

build_browser_hub(
  these_samples_metadata = my_meta,
  these_seq_types = c("genome", "capture"),
  projection = "grch37",
  local_web_host_dir = local_web_host_dir,
  hub_dir = hub_dir,
  splitColumnName = "pathology",
  longLabel = "Public aSHM mutations separated by pathologies",
  contact_email = "rdmorin@sfu.ca"
)
}

}
