% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleanup_maf.R
\name{cleanup_maf}
\alias{cleanup_maf}
\title{Cleanup MAF.}
\usage{
cleanup_maf(maf_df)
}
\arguments{
\item{maf_df}{input MAF data frame.}
}
\value{
maf_df with transformed columns
}
\description{
Transform input maf columns to allow for usage of dplyr verbs.
}
\details{
Transform input maf columns to allow for usage of dplyr verbs.
Allowing for a stright-forward plotting workflow as well as downstream data aggregation and manipulation.
This function expects a set number of columns to exist in the incoming maf in order for this to work.
To view the columns, see bundled data in GAMBLR.data.
}
\examples{

clean_maf = cleanup_maf(maf_df = GAMBLR.data::sample_data$grch37$maf)

}
