% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cn_core_processor.R
\name{cn_core_processor}
\alias{cn_core_processor}
\title{Copy Number (CN) Core Processor}
\usage{
cn_core_processor(
  cn_state_matrix,
  remove_chrom = c("chrX"),
  round_first = FALSE,
  drop_bin_if_sd_below = 0,
  nrm_by_avg_ploidy = FALSE,
  simplyfy_nrm = TRUE,
  auto_calc_cutoff = FALSE,
  auto_calc_cutoff_method = "median",
  cutoff_gain_custom = 0.05,
  cutoff_loss_custom = 0.005
)
}
\arguments{
\item{cn_state_matrix}{A matrix of copy number (CN) states exported from get_cn_states function, where rows represent samples and columns represent genomic bins (segments).}

\item{remove_chrom}{A character vector of chromosomes to exclude from analysis (e.g., "chrX"). Default is \code{"chrX"}.}

\item{round_first}{Logical. If \code{TRUE}, CN values are rounded to integers before further analysis. Default is \code{FALSE}.}

\item{drop_bin_if_sd_below}{Numeric. Bins (segments) with a standard deviation below this threshold are excluded. Default is \code{0}.}

\item{nrm_by_avg_ploidy}{Logical. If \code{TRUE}, CN values are normalized by the average ploidy across all samples. Default is \code{FALSE}.}

\item{simplyfy_nrm}{Logical. If \code{TRUE}, normalized CN values are rounded to integers for simplicity. Default is \code{TRUE}.}

\item{auto_calc_cutoff}{Logical. If \code{TRUE}, cutoff values for CN state classification are calculated automatically using the method specified in \code{auto_calc_cutoff_method}. Default is \code{FALSE}.}

\item{auto_calc_cutoff_method}{Character. Specifies the method for automatic cutoff calculation. Options are \code{"median"}, \code{"1st_quartile"}, \code{"3rd_quartile"}, or \code{"mean"}. Default is \code{"median"}.}

\item{cutoff_gain_custom}{Numeric. Custom cutoff value for classifying gains in CN states, used when \code{auto_calc_cutoff} is \code{FALSE}. Default is \code{0.05}.}

\item{cutoff_loss_custom}{Numeric. Custom cutoff value for classifying losses in CN states, used when \code{auto_calc_cutoff} is \code{FALSE}. Default is \code{0.005}.}
}
\value{
An object of S4 class \code{all} containing:
\itemize{
\item \code{mat}: A \code{mat} object with filtered CN state matrix and optionally normalized matrix.
\item \code{data}: A \code{data} object containing filtered bin data, CNAR statistics for bins and lengths.
}
}
\description{
The primary purpose is to extract statistics from \code{pretty_CN_heatmap} with customized modifications, such as filtering and normalization of copy number (CN) state data and calculating Copy Number Alteration Ratios (CNARs).
}
\examples{
metadata = get_gambl_metadata() \%>\% filter(pathology == "DLBCL")
cn_state_matrix = get_cn_states(n_bins_split=5000,
                                missing_data_as_diploid = TRUE, # if FALSE, missing data will be NA
                                seg_data = all_segments,
                                these_samples_metadata = metadata)
result = cn_core_processor(cn_state_matrix, 
                           metadata, 
                           remove_chrom = c("chrX"), 
                           auto_calc_cutoff=FALSE,
                           cutoff_gain_custom = 0.05,
                           cutoff_loss_custom = 0.005)
}
