% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genomic_data.R
\name{dplyr_reconstruct.genomic_data}
\alias{dplyr_reconstruct.genomic_data}
\title{Reconstruct genomic_data objects after dplyr operations}
\usage{
\method{dplyr_reconstruct}{genomic_data}(data, template)
}
\arguments{
\item{data}{A data frame resulting from a dplyr operation.}

\item{template}{The original \code{genomic_data} object used as a template.}
}
\value{
A \code{genomic_data} object with preserved attributes.
}
\description{
This method is invoked by dplyr after performing operations such as
\code{mutate}, \code{filter}, \code{select}, etc. It ensures that custom
attributes (e.g., \code{genome_build}) and the custom class order for objects
of class \code{genomic_data} are preserved.
}
