% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GAMBLR.utils-package.R
\docType{package}
\name{GAMBLR.utils-package}
\alias{GAMBLR.utils}
\alias{GAMBLR.utils-package}
\title{GAMBLR.utils: Collection of functions that consist of the core functionality of GAMBLR}
\description{
This package is part of the Genomic Analysis of Mature B-cell Lymphomas (GAMBL) project developed by the Morin Lab.
}
\author{
\strong{Maintainer}: Vladimir Souza \email{vsouza@bcgsc.ca}

}
\keyword{internal}
