% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reannotate_ssm.R
\name{reannotate_ssm}
\alias{reannotate_ssm}
\title{Reannotate SSMs in a MAF format}
\usage{
reannotate_ssm(maf_data, preferred_transcripts)
}
\value{
A data frame in maf_data format
}
\description{
Pick another annotation for SSMs annotated by VEP using the additional isoforms
}
\details{
To Do
}
\examples{
\dontrun{
csnk2b_region_maf = get_ssm_by_regions(regions_list="chr6:31665606-31674128",
                                     these_samples_metadata = get_gambl_metadata(),
                                      projection="hg38",
                                      basic_columns=F,streamlined=F,this_seq_type = "genome")
csnk2b_region_maf_cap = get_ssm_by_regions(regions_list="chr6:31665606-31674128",
                                     these_samples_metadata = get_gambl_metadata(),
                                      projection="hg38",
                                      basic_columns=F,streamlined=F,this_seq_type = "capture")
}
}
