% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_cn_matrix_by_genes.R
\name{annotate_cn_matrix_by_genes}
\alias{annotate_cn_matrix_by_genes}
\title{Annotate copy number matrix by gene names}
\usage{
annotate_cn_matrix_by_genes(these_samples_metadata, gene_ids, cn_matrix)
}
\arguments{
\item{these_samples_metadata}{The metadata for samples of interest to be included in the returned matrix.
Can be created with \code{get_gambl_metadata} function.}

\item{gene_ids}{The "gene_id" column stores gene symbols (characters).}

\item{cn_matrix}{User can provide a matrix of CN values for the samples in the metadata.
See \link{segmented_data_to_cn_matrix} for more information on how to create this matrix.}
}
\value{
A copy number matrix with sample ids as rows and gene names as columns
}
\description{
Annotates a copy number matrix using user-provided gene names
}
\examples{

\dontrun{
# Get sample metadata 
dlbcl_meta = get_gambl_metadata() \%>\% filter(pathology == "DLBCL")

# Consider some genes
gene_symbols = c("MYC", "MIR17HG", "CCND3","ID3","DDX3X", "SYNCRIP")

# Get cn_matrix
all_segments = get_cn_segments(these_samples_metadata = dlbcl_meta)
cn_matrix = segmented_data_to_cn_matrix(
   seg_data = all_segments,
    strategy="auto_split",
   these_samples_metadata = dlbcl_meta,
   adjust_for_ploidy=TRUE
)

annotate_cn_matrix_by_genes (these_samples_metadata = dlbcl_meta,
                             gene_ids = gene_symbols,
                             cn_matrix = cn_matrix
)
}
}
