% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate_curated_hotspots.R
\name{annotate_curated_drivers}
\alias{annotate_curated_drivers}
\title{Annotate curated drivers}
\usage{
annotate_curated_drivers(
  maf_data,
  genes_of_interest = c("MYD88", "NOTCH1", "NOTCH2"),
  genome_build,
  custom_coordinates,
  existing_values_action = "clobber"
)
}
\arguments{
\item{genes_of_interest}{An optional vector of genes for hotspot annotation. By default, the genes present in the curated hotspot tables are supported.}

\item{genome_build}{Reference genome build for the coordinates in the MAF file. The default is grch37 genome build.}

\item{custom_coordinates}{A data frame with custom coordinates for the hot spots.
All mutations in any of the regions specified in the data frame will be marked as hot spots.
The data frame must have the following columns: "Hugo_Symbol", "chrom", "start", and "end".
Optional columns are: "classes" (regex for Variant_Classification) and "alias" (identifier for the hotspot).
The "type" column can include tokens such as missense, inframe, trunc, startloss,
stoploss, splicesite, and spliceregion (comma-separated).}

\item{existing_values_action}{Character. How to handle existing columns.
Use "clobber" (default) to always reset "hot_spot" and "mutation_alias"
before re-annotating. Use "update" to only fill missing values.}

\item{annotated_maf}{A maf_data object or data frame in MAF format that has hotspots annotated using the function annotate_hotspots().}
}
\value{
The same data frame (as given to the \code{annotated_maf} parameter) with the reviewed columns
"hot_spot", "mutation_alias", and "mutation_annotation".
}
\description{
Annotate MAF-like data frame with a hot_spot column indicating recurrent mutations.
}
\details{
This function annotates a MAF-like data frame and will create or overwrite the
"hot_spot" column based on curated hotspot rules. Genes for hotspot review are supplied
with the \code{genes_of_interest} parameter. Overlap is determined using a fuzzy interval
match so any overlap between a mutation and a hotspot region is considered a match.
Currently only a few sets of genes are supported, see parameter description for more information and limitations.
The desired genome build can be specified with \code{genome_build} parameter (useful if you loaded the MAF from disk).
Obviously, you need to specify the same genome_build as the coordinate system used in the incoming MAF.
}
\examples{

# Get metadata for all DLBCLs available
dlbcl_meta = GAMBLR.open::get_gambl_metadata() \%>\% 
   dplyr::filter(pathology == "DLBCL")
# get mutations (as maf_data) for all these samples including non-coding (silent) for DLBCLone
dlbcl_coding = GAMBLR.open::get_all_coding_ssm(
                                   dlbcl_meta,
                                   include_silent=TRUE)

# annotate known driver mutations using bundled annotations
# WARNING: currently hg38 annotations are out of sync with grch37
# Stick with grch37 for now
dlbcl_coding = annotate_curated_drivers(maf_data=dlbcl_coding)

#look at the new columns we have
dplyr::filter(dlbcl_coding,
             !grepl("other",mutation_alias)) \%>\%
 dplyr::select(Chromosome,Start_Position,Variant_Classification,HGVSp_Short,hot_spot,mutation_alias,mutation_annotation)


}
