% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_ssm_by_region.R
\name{count_ssm_by_region}
\alias{count_ssm_by_region}
\title{Count SSM In A Region}
\usage{
count_ssm_by_region(
  region,
  chromosome,
  start,
  end,
  all_mutations_in_these_regions,
  these_samples_metadata,
  count_by,
  seq_type = "genome"
)
}
\arguments{
\item{region}{Region formatted like chrX:1234-5678 instead of specifying chromosome, start and end separately.}

\item{chromosome}{The chromosome you are restricting to (with or without a chr prefix).}

\item{start}{Query start coordinate of the range you are restricting to.}

\item{end}{Query end coordinate of the range you are restricting to.}

\item{all_mutations_in_these_regions}{If you are calling this function many times (e.g. bins spanning a larger region), to save a ton of time you are strongly encouraged to provide the output of \code{get_ssm_by_region} on the entire region of interest and passing it to this function}

\item{these_samples_metadata}{A metadata table subset to the sample IDs of interest. If not provided, the function will return all metadata and regions will be returned for all samples in the metadata.}

\item{count_by}{Defaults to counting all variants. Specify 'sample_id' if you want to collapse and count only one per sample}

\item{seq_type}{The seq_type you want back, default is genome.}
}
\description{
Count the variants in a region with a variety of filtering options.
}
\examples{
library(GAMBLR.data)

#define a region.
my_region = gene_to_region(gene_symbol = "MYC", return_as = "region")

#count SSMs for the selected sample subset and defined region.
count_ssm_by_region(region = my_region)

}
