% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/liftover.R
\name{liftover}
\alias{liftover}
\title{Liftover.}
\usage{
liftover(
  data_df,
  mode = "bedpe",
  target_build = "grch37",
  bedpe_file,
  col_names,
  col_types,
  standard_bed = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{data_df}{Input data to be lifted to another projection in the format of data frame.}

\item{mode}{Specify the type of data to be converted. Available options are "bedpe" (default), "maf", and "bed".}

\item{target_build}{Specify which build the data should be lifted to (must be one of hg19, grch37, hg38, grch38).}

\item{bedpe_file}{Optionally, specify the path to a bedpe file if it is not already loaded to the environment. Only used in the bedpe mode.}

\item{col_names}{If not provided, the column names will be imposed. Only used in the bedpe mode.}

\item{col_types}{Specify column types if column names are also defined with \code{col_names}. Only used in the bedpe mode.}

\item{standard_bed}{Boolean parameter for defining the type of bed file that is provided with \code{bedpe_file}. Deafult is FALSE. Only used in the bedpe mode.}

\item{verbose}{Set to TRUE for verbose output. Default is FALSE.}
}
\value{
Data frame containing original bedpe data with new coordinates.
}
\description{
Use liftOver to convert a coordinate-based file between the two main genome builds (grch37/hg38).
}
\details{
The user can specify the data to be converted to another genome build in a data frame with \code{data_df}.
For the bedpe data, the user can optionally also specify a path to the bedpe file that needs to be lifted with \code{bedpe_file}
(remained here for backwards compatibility).

The other required parameter is \code{target_build}, this parameter decides the final projection of the lifted data.

The data type can be specified with the \code{mode} argument, which accepts one of the bedpe (default, maf, or bed files).
Other formats are not supported yet.
}
\examples{
bed <- grch37_ashm_regions

test <- liftover(
    data_df = bed,
    mode = "bed",
    target_build = "hg38"
)

maf <- get_coding_ssm(
     these_sample_ids = "DOHH-2"
)

test <- liftover(
    data_df = maf,
    mode = "maf",
    target_build = "hg38"
)

bedpe <- get_manta_sv(
     these_sample_ids = "DOHH-2"
)

test <- liftover(
    data_df = bedpe,
    mode = "bedpe",
    target_build = "hg38"
)

}
