% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_region_to_bins.R
\name{map_regions_to_bins}
\alias{map_regions_to_bins}
\title{Map region to bins}
\usage{
map_regions_to_bins(
  query_regions,
  regions,
  query_type = "region",
  first = TRUE
)
}
\arguments{
\item{first}{}
}
\value{
a named list
}
\description{
Map region to bins
}
\examples{

gene_region <- gene_to_region("TP53")

all_bins <- colnames(cn_state_matrix)

TP53_bin <- map_region_to_bins(gene_region, all_bins, first = TRUE)

}
