% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmented_data_to_cn_matrix.R
\name{process_cn_segments_by_region}
\alias{process_cn_segments_by_region}
\title{Filter and summarise CN values seg_data for a single genomic region}
\usage{
process_cn_segments_by_region(
  seg_data,
  region,
  streamlined,
  weighted_average = TRUE
)
}
\arguments{
\item{seg_data}{}

\item{region}{}

\item{streamlined}{}

\item{weighted_average}{}
}
\value{
data.frame
}
\description{
Filter and summarise CN values seg_data for a single genomic region
}
