% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmented_data_to_cn_matrix.R
\name{process_cn_segments_by_region}
\alias{process_cn_segments_by_region}
\title{Filter and summarise CN values seg_data for a single genomic region}
\usage{
process_cn_segments_by_region(
  seg_data,
  region,
  streamlined = FALSE,
  weighted_average = TRUE,
  fill_missing_with = "avg_ploidy"
)
}
\arguments{
\item{seg_data}{Data that will be processed for the specified region}

\item{region}{Region to be processed in the format "chromosome:start-end"}

\item{streamlined}{If TRUE, only return the ID and CN columns (default is FALSE)}

\item{weighted_average}{If TRUE, calculate the CN value as a weighted average of the segments
that overlap the region. Otherwise, use the CN value of the first segment that overlaps the region (default is TRUE)}

\item{fill_missing_with}{Specify how the value will be assigned to any region not covered by a segment: Options: "diploid" or "avg_ploidy"}
}
\value{
data.frame
}
\description{
Assign the copy number value to a region based on the segment(s) that overlap it
}
\details{
This function returns CN states for a single region specified in the format "chromosome:start-end"
using segmented copy number data (seg_data). It will either return a segment in the standard format (all original columns)
or a streamlined format with only the ID and CN columns
}
\examples{

region_segs = process_cn_segments_by_region(
       region = "chrX:1-1000000",
       streamlined = FALSE, 
       weighted_average = T, 
       seg_data = seg_data)

}
