% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_maf.R
\name{supplement_maf}
\alias{supplement_maf}
\title{Supplement MAF.}
\usage{
supplement_maf(incoming_maf, these_samples_metadata)
}
\arguments{
\item{incoming_maf}{The initial MAF data frame to be supplemented with missing samples.}

\item{these_samples_metadata}{The metadata data frame that contains Tumor_Sample_Barcode column with ids to be present in the complemented MAF.}
}
\value{
maf_df with complemented Tumor_Sample_Barcode and other columns ready to be used downstream.
}
\description{
Complement maf with missing samples.
}
\details{
Specify the initial MAF with \code{incoming_maf} (to be supplemented with missing samples) and
give the function a filtered metadata table (with the sample IDs of interest) to the \code{these_samples_metadata}.
}
\examples{
my_metadata = GAMBLR.data::gambl_metadata
reddy_meta = dplyr::filter(my_metadata, cohort=="dlbcl_reddy")

small_maf = dplyr::filter(GAMBLR.data::sample_data$grch37$maf,
 Tumor_Sample_Barcode \%in\% reddy_meta$Tumor_Sample_Barcode)

small_maf = dplyr::filter(small_maf, Hugo_Symbol == "MYC")

complete_maf = supplement_maf(incoming_maf = small_maf,
                              these_samples_metadata = reddy_meta)

}
